python-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
scipy==1.11.4
openpyxl==3.1.2
xlrd==2.0.1
pymongo==4.6.1
openai==1.10.0
google-generativeai==0.3.2
requests==2.31.0
